fixing js issues
