+++ package.json
@@ -30,7 +30,9 @@
   "dependencies": {
     "chart.js": "^2.7.3",
     "electron-squirrel-startup": "^1.0.0",
-    "jquery": "^3.3.1"
+    "java": "^0.11.0",
+    "jquery": "^3.3.1",
+    "node-java-maven": "^0.1.1"
   },
   "description": "Spline generator for monkeys.",
   "devDependencies": {
@@ -41,6 +43,7 @@
     "@electron-forge/maker-zip": "^6.0.0-beta.30",
     "assert": "^1.4.1",
     "electron": "3.0.10",
+    "electron-rebuild": "^1.8.2",
     "mocha": "^5.2.0"
   },
   "keywords": [],
